% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abind3.R
\name{abind3}
\alias{abind3}
\title{Convenience function for stacking matrices into an array.}
\usage{
abind3(...)
}
\arguments{
\item{...}{Any number of matrices of equal dimension to stack together into
a 3d matrix}
}
\description{
This function extends the abind function from the abind package.
}
