% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayToLong.R
\name{arrayToLong}
\alias{arrayToLong}
\title{Reshape array of data matrices into long format}
\usage{
arrayToLong(X, coords, yrs)
}
\arguments{
\item{X}{3 dimensional array of matrices to extract to long format}

\item{coords}{Spatial coordinates associated with the data (longitude in first column)}

\item{yrs}{Vector with labels for the years}
}
\description{
Reshape array of data matrices into long format
}
