% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStData.R
\name{buildStData}
\alias{buildStData}
\title{Assemble stData object from raw components}
\usage{
buildStData(X, Y, Z, tLabs, coords.r, coords.s = NULL, Q = NULL,
  X.lab = "", Y.lab = "", Z.lab = "")
}
\arguments{
\item{X}{3-dimensional array with \eqn{n} rows, \eqn{p} columns, and \eqn{t}
slices. \code{X[i,j,k]} represents the \code{j}th covariate observed at 
time \code{k} at location \code{i}.}

\item{Y}{matrix with \eqn{n} rows and \eqn{t} columns. \code{Y[i,j]} 
represents the \code{j}th response observed at location \code{i}.}

\item{tLabs}{vector of labels for each timepoint}

\item{coords.r}{matrix with coordinates where remote covariates
were observed (lon, lat)}

\item{coords.s}{(optional) matrix with coordinates where responses were 
observed (lon, lat).  If \code{coords.s} is not provided, then \code{Q}
must be provided}

\item{Q}{if spatial locations for \code{X} and \code{Y} represent areal 
units, then \code{Q} represents the adjacency matrix for the spatial 
configuration}

\item{X.lab}{name for X data (optional)}

\item{Y.lab}{name for Y data (optional)}

\item{Z.lab}{name for Z data (optional)}
}
\description{
\code{telefit} fitting and prediction objects rely on \code{stData} objects,
which are \code{list} objects with some common structure.  \code{buildStData}
is a convenience function for creating the list object that also adds some 
basic validation that input objects have the right dimensions, etc.
}
\examples{
n = 100
p = 3
t = 10
nz = 300

x = buildStData(X = array(rnorm(n = n*p*t), dim = c(n,p,t)), 
                Y = matrix(rnorm(n = n*t), nrow = n), 
                Z = matrix(rnorm(n = nz*t), nrow = nz), 
                tLabs = 1:t, 
                coords.r = matrix(rnorm(n = nz*2), ncol = 2),
                coords.s = matrix(rnorm(n = n*2), ncol = 2), 
                Q = NULL, 
                X.lab = 'xlab', Y.lab = 'ylab', Z.lab = 'zlab')
}
