% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caprecip.counts}
\alias{caprecip.counts}
\title{Counts of large storms in CA}
\format{A stData object with 33 years of observations}
\source{
\url{http://prism.oregonstate.edu}

\url{https://rda.ucar.edu/datasets/ds627.0/}
}
\usage{
caprecip.counts
}
\description{
A dataset containing sample spatially-aggregated climate data from the 
ERA-Interim and PRISM datasets.  The response comes from PRISM and 
represents the number of large storms that occured within a HUC8 subbasin
during California DJF winters.  The covariates come from ERA-Interim, 
Colorado and Pacific Ocean (sea) surface temperatures; covariates are also 
available that describe the size of the subbasins.
}
\examples{

data("caprecip.counts")
str(caprecip.counts)

}
\keyword{datasets}
