% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca.R
\name{cca}
\alias{cca}
\title{Compute canonical correlations between multivariate datasets using EOF decompositions}
\usage{
cca(X, Y, X.k, Y.k)
}
\arguments{
\item{X}{An \code{(nvars x nobs)} data frame or matrix in which each column contains 
all observations of measured (predictor) variables for a given timepoint or
sample.  For example, if X represents a spatial variable that was recorded 
at several timepoints, then each row of X should contain the variable's 
measurement for all timepoints at a single location.}

\item{Y}{An \code{(nresponse x nobs)} data frame or matrix in which each
column contains all observations of measured (response) variables for a 
given timepoint or sample.  For example, if Y represents a spatial variable 
that  was recorded  at several timepoints, then each row of Y should contain 
the variable's  measurement for all timepoints at a single location.}

\item{X.k}{Number of empirical orthogonal functions (i.e., principal 
components) from the X data to use in the analysis.}

\item{Y.k}{Number of empirical orthogonal functions (i.e., principal 
components) from the Y data to use in the analysis.}
}
\value{
A list containing the following components:
 \describe{
   \item{\code{X.eof}}{EOF decomposition of X.}
   \item{\code{Y.eof}}{EOF decomposition of Y.}
   \item{\code{X.k}}{Number of EOFs of X used in analysis.}
   \item{\code{Y.k}}{Number of EOFs of Y used in analysis.}
   \item{\code{cor}}{Correlations of canonical variables.}
   \item{\code{A}}{Canonical patterns for X in EOF-space. Each column 
     represents one pattern.}
   \item{\code{B}}{Canonical patterns for Y in EOF-space. Each column 
     represents one pattern.}
   \item{\code{C}}{Time series of canonical amplitudes (a.k.a. scores) for X. 
     Each column represents the amplitude of one pattern over time.}
   \item{\code{D}}{Time series of canonical amplitudes (a.k.a. scores) for Y. 
     Each column represents the amplitude of one pattern over time.}
   \item{\code{F.X}}{Canonical patterns for X in the same coordinate space 
     as X. Each column represents one pattern.}
   \item{\code{F.Y}}{Canonical patterns for Y in the same coordinate space 
     as Y. Each column represents one pattern.}
 } 
 
 See Cook, et. al. (1994) for details about objects A, B, C, and D; see 
 von Storch and Zwiers (1999) for details about F.X and F.Y.
}
\description{
Based on Glahn (1968) and Cook, et. al. (1994).
}
\details{
This will center and scale data for efficiency.
}
\references{
Cook, E.R., Briffa, K.R., and Jones, P.D., 1994, Spatial regression methods in dendroclimatology: A review and comparison of two techniques: International Journal of Climatology, v. 14, p. 379–402.

Glahn, H.R., 1968, Canonical Correlation and Its Relationship to Discriminant Analysis and Multiple Regression: Journal of the Atmospheric Sciences, v. 25, p. 23–31, doi: 10.1175/1520-0469(1968)025<0023:CCAIRT>2.0.CO;2.

Storch, H. Von, and Zwiers, F.W., 1999, Statistical Analysis in Climate Research: Cambridge, Cambrudge University Press.
}
\seealso{
\link{eof} For details about EOF decompositions.
}
