% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coprecip.fit}
\alias{coprecip.fit}
\title{Sample MCMC output for the RESP model}
\format{An stFit object, which is a list of several objects
\describe{
  \item{parameters}{MCMC samples of model parameters}
  \item{priors}{description of priors used to fit model}
  \item{miles}{TRUE or FALSE to specify whether the spatial distances 
      used to estimate spatial covariance parameters were in units of miles 
      (TRUE) or kilometers (FALSE)}
  \item{localOnly}{TRUE if remote covariates were not estimated}
  \item{remoteOnly}{TRUE if local covariates were not estimated}
  \item{varying}{(deprecated) TRUE if local covariates were estimated as a 
      spatially-varying field}
  \item{coords.knots}{coordinates of remote knot locations}
}}
\usage{
coprecip.fit
}
\description{
An example stFit object containing output from a short run of the MCMC 
sampler that fits the RESP model to data.
}
\examples{

data("coprecip.fit")
str(coprecip.fit)

}
\keyword{datasets}
