% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.tel.R
\name{cor.tel}
\alias{cor.tel}
\title{Compute pointwise correlations for an exploratory teleconnection analysis}
\usage{
cor.tel(Y, Z, coords.local, coords.remote, ncores = 1, wide = F,
  wide.label.signif = 5)
}
\arguments{
\item{Y}{[ny x nt]}

\item{Z}{[nz x nt]}

\item{coords.local}{coordinates of locations in Y}

\item{coords.remote}{coordinates of locations in Z}

\item{wide}{if TRUE, then the results will be coerced into a matrix}
}
\description{
Compute pointwise correlations for an exploratory teleconnection analysis
}

