% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgemkmm.R
\name{dgemkmm}
\alias{dgemkmm}
\title{Evaluate kron(A,B) * C without storing kron(A,B)}
\usage{
dgemkmm(A, B, C)
}
\arguments{
\item{A}{(m x n) matrix}

\item{B}{(p x q) matrix}

\item{C}{(nq x r) matrix}
}
\description{
Evaluate kron(A,B) * C without storing kron(A,B)
}
