% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eof.R
\name{eof}
\alias{eof}
\title{Performs an EOF decomposition of the data}
\usage{
eof(Z, center = T, scale = T)
}
\arguments{
\item{Z}{A spatio-temporal observation matrix of a field for which spatial 
EOFs will be computed.  Each column should have all spatial observations
for a given timepoint.}
}
\value{
A list with the following components:  
 
 patterns (columns of Phi)
 amplitudes (sc)
 cumvar (cumulative proportion of variance explained by each pattern)
 mu (climatology)
}
\description{
Based on Wikle's text
}
\details{
Each column in the amplitude contains the time series for one EOF pattern
Each column in the pattern contains the spatial loadings for one EOF
}

