% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errDump.R
\name{errDump}
\alias{errDump}
\title{Wrapper for a function to dump errors from C++}
\usage{
errDump(x, fname = file.path(tempdir(), "error_samplerState.RData"))
}
\arguments{
\item{x}{Data to save}

\item{fname}{Path/name to save data to}
}
\description{
Wrapper for a function to dump errors from C++
}
