% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAnomalies.R
\name{extractAnomalies}
\alias{extractAnomalies}
\title{Compute anomalies for a SpatialGridDataFrame and returns an extent}
\usage{
extractAnomalies(sgdf, extent, aggfact = NULL, mask = NULL)
}
\arguments{
\item{extent}{raster::extent object featuring region to extract}

\item{aggfact}{if provided, will spatially average the data}

\item{mask}{if an sgdf is provided, the data will be masked before
extraction, aggregation, and anomaly computation}
}
\value{
a modified SpatialGridDataFrame, sgdf, with the climatology for each
 location accessible via attr(sgdf@data@values, 'scaled:center')
}
\description{
Compute anomalies for a SpatialGridDataFrame and returns an extent
}
\examples{
# Show an example where you plot the climatology grid too!


}

