% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDinterval.stFit.R
\name{HPDinterval.stFit}
\alias{HPDinterval.stFit}
\title{Compute Highest posterior density intervals from posterior samples}
\usage{
\method{HPDinterval}{stFit}(stFit, burn = 1, prob = 0.95)
}
\arguments{
\item{stFit}{stFit object containing posterior samples for model}

\item{burn}{number of posterior samples to reject before computing estimates}

\item{prob}{The target probability content of the intervals}
}
\description{
Compute Highest posterior density intervals from posterior samples
}
\examples{

data("coprecip.fit")
HPDinterval.stFit(coprecip.fit, burn = 50)

}
