% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invWSamp.R
\name{invWSamp}
\alias{invWSamp}
\title{Samples an Inverse-Wishart matrix}
\usage{
invWSamp(Psi, n)
}
\arguments{
\item{Psi}{an \eqn{n x n} scale matrix}

\item{n}{degrees of freedom parameter}
}
\description{
Samples \eqn{W ~ IW(Psi, n)}
}
\examples{

A = matrix(c(1,.5,.5,1), ncol=2)
W = invWSamp(A, 3)

}
