% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronSamp.R
\name{kronSamp}
\alias{kronSamp}
\title{Samples a multivariate normal with a Kronecker product covariance structure}
\usage{
kronSamp(A, B)
}
\arguments{
\item{A}{an \eqn{m x n} matrix}

\item{B}{an \eqn{p x q} matrix}
}
\description{
Samples \eqn{x ~ N(0, AxB)}
}
\examples{

A = matrix(c(1,.5,.5,1), ncol=2)
B = diag(2)
x = kronSamp(A, B)

}
