% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maternArray.R
\name{maternArray}
\alias{maternArray}
\title{Matern covariance}
\usage{
maternArray(d, scale = 1, range = 1, smoothness = 0.5, nugget = 0)
}
\arguments{
\item{d}{A numeric vector of distances at which the Matern 
correlation function should be evaluated.}

\item{scale}{Scales correlations to covariances.}

\item{range}{Matern range parameter.  Controls the decay of pointwise 
correlations as a function of distance.}

\item{smoothness}{Matern smoothness parameter.  Controls the number of 
process derivatives.}

\item{nugget}{Spatial covariance nugget.}
}
\description{
This function evaluates the Matern covariance function for the elements of 
a vector.
}
