% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maternCov.R
\name{maternCov}
\alias{maternCov}
\title{Matern covariance}
\usage{
maternCov(d, scale = 1, range = 1, smoothness = 0.5, nugget = 0)
}
\arguments{
\item{d}{A numeric vector or matrix of distances at which the Matern 
correlation function should be evaluated.}

\item{scale}{Scales correlations to covariances.}

\item{range}{Matern range parameter.  Controls the decay of pointwise 
correlations as a function of distance.}

\item{smoothness}{Matern smoothness parameter.  Controls the number of 
process derivatives.}

\item{nugget}{Spatial covariance nugget.}
}
\description{
This function evaluates the Matern covariance function for the elements of 
a spatial distance matrix, d---i.e., a symmetric matrix with 0's along its
main diagonal.
}

