% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeComposition.R
\name{mergeComposition}
\alias{mergeComposition}
\title{Combine results from composition sampler}
\usage{
mergeComposition(xfull, yfull)
}
\arguments{
\item{xfull}{Raw output from one run of the Rcpp/Armadillo composition sampler}

\item{yfull}{Raw output from another run of the Rcpp/Armadillo composition sampler}
}
\description{
Combine results from composition sampler
}
