% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMean.R
\name{mergeMean}
\alias{mergeMean}
\title{Combine sample means from two samples}
\usage{
mergeMean(x.mean, y.mean, x.n, y.n)
}
\arguments{
\item{x.mean}{sample mean from the first sample, 'x'}

\item{y.mean}{sample mean from the second sample, 'y'}

\item{x.n}{sample size from the first sample, 'x'}

\item{y.n}{sample size from the second sample, 'y'}
}
\description{
This function combines the sample mean information from two samples (of 
the same phenomena) to return the sample mean of the union of the two 
samples.
}
