% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairsplot.R
\name{pairsplot}
\alias{pairsplot}
\title{Creates a pairwise plot of the sampled parameters}
\usage{
pairsplot(stFit, burn = 1)
}
\arguments{
\item{stFit}{return object from telefit::stFit}

\item{burn}{number of samples to discard before plotting}
}
\description{
Creates a pairwise plot of the sampled parameters
}

