% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cor.tel.R
\name{plot.cor.tel}
\alias{plot.cor.tel}
\title{Plots teleconnection correlation maps}
\usage{
plot.cor.tel(cor.tel, coord.Y, lon.W, lon.E, lat.S, lat.N, signif = F,
  boxsize = NULL)
}
\arguments{
\item{cor.tel}{an output object from cor.tel}

\item{coord.Y}{c(lon, lat) specifying a base point for which to plot remote 
correlations}

\item{lon.W}{westmost longitude to plot}

\item{lon.E}{eastmost longitude to plot}

\item{lat.S}{southmost latitude to plot}

\item{lat.N}{northmost latitude to plot}

\item{signif}{if TRUE, then cor.tel must have a column labeled 'signif' that
indicates which correlations are significant.  These correlations will be
printed in bold, and the rest will be printed more lightly}

\item{boxsize}{size of grid boxes plotted}
}
\value{
a ggplot object with the teleconnection map
}
\description{
This function provides basic plotting for analyses returned from cor.tel
}

