% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.density.R
\name{plot.density}
\alias{plot.density}
\title{Creates a density plot of the sampled parameters}
\usage{
plot.density(stFit, burn = 1, text.size = 36, axis.text.size = 24)
}
\arguments{
\item{stFit}{return object from telefit::stFit}

\item{burn}{number of samples to discard before plotting}
}
\value{
a ggplot object with the density plots
}
\description{
Creates a density plot of the sampled parameters
}

