% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.response.R
\name{plot.response}
\alias{plot.response}
\title{Plots teleconnection correlation maps}
\usage{
\method{plot}{response}(Y, boxsize = NULL)
}
\arguments{
\item{Y}{a response field to plot; must be in long format with (lon.Y, lat.Y)
coordinates}

\item{boxsize}{size of grid boxes plotted}
}
\value{
a ggplot object with the response map
}
\description{
This function provides basic plotting for analyses returned from cor.tel
}

