% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeCheck.R
\name{rangeCheck}
\alias{rangeCheck}
\title{Determine basic interval over which a covariance range parameter makes numerically
invertible matrices}
\usage{
rangeCheck(coords.local, coords.remote, Z, miles = T, smoothness.local = 2,
  smoothness.remote = 1.76, nugget.local = 0, scale.remote = 1,
  maxIt = 1000, tol = 1e-10)
}
\description{
Note that the remote range numbers don't necessarily mean anything since the
full model is quite generous.  All the same, it's informative.
}
