% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmatnorm.R
\name{rmatnorm}
\alias{rmatnorm}
\title{Simulate matrices from matrix normal distributions}
\usage{
rmatnorm(n, U, V, M = matrix(0, nrow = nrow(U), ncol = nrow(V)))
}
\arguments{
\item{n}{Number of random matrices to simulate}

\item{U}{Covariance matrix defining dependence between rows}

\item{V}{Covariance matrix defining dependence between columns}

\item{M}{average value of each entry in the sampled matrices}
}
\description{
Draw random matrices from the matrix normal distribution \deqn{MN(M, U, V)}  
Note that an observation, \eqn{X}, from this equation has the following 
distribution when vectorized \deqn{vec(X) ~ N(vec(M), kron(V, U) )}
}
