% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwishart.R
\name{rwishart}
\alias{rwishart}
\title{Random wishart matrix}
\usage{
rwishart(V, n)
}
\arguments{
\item{V}{symmetric positive definite p x p scale matrix}

\item{n}{degrees of freedom (greater than p-1)}
}
\description{
Random wishart matrix
}
