% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stRecoverAlpha.R
\name{stRecoverAlpha}
\alias{stRecoverAlpha}
\title{Matern covariance}
\usage{
stRecoverAlpha(stFit, X, Y, Z, coords.local, coords.remote, nu_y, nu_r, burn,
  prob = 0.95, miles = T, ncores = 1, summaryOnly = T)
}
\arguments{
\item{summaryOnly}{Only return alpha estimates and key variances; do not 
return all posterior samples.}
}
\description{
This function evaluates the Matern covariance function for the elements of 
a spatial distance matrix, d---i.e., a symmetric matrix with 0's along its
main diagonal.
}

