% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stSimulate.R
\name{stSimulate}
\alias{stSimulate}
\title{Compute forecasts based on posterior samples}
\usage{
stSimulate(stFit, coords.local, X, Z, yearLabs, nu_y, miles = T, ncores = 1)
}
\description{
Compute forecasts based on posterior samples
}

