% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telefit.R
\docType{package}
\name{telefit}
\alias{telefit}
\title{Tools for modeling teleconnections}
\description{
The package \pkg{telefit} provides functions for fitting the remote 
effects spatial process (RESP) model.
}
