% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot.R
\name{traceplot}
\alias{traceplot}
\title{Creates a traceplot of the sampled parameters}
\usage{
traceplot(stFit, burn = 1, text.size = 36, axis.text.size = 24)
}
\arguments{
\item{stFit}{return object from telefit::stFit}

\item{burn}{number of samples to discard before plotting}
}
\value{
a ggplot object with the traceplot
}
\description{
Creates a traceplot of the sampled parameters
}

