% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca.predict.R
\name{cca.predict}
\alias{cca.predict}
\title{Make predictions using canonical correlation analysis results}
\usage{
cca.predict(obj, X)
}
\arguments{
\item{obj}{\code{cca} object containing results of a cca analysis.}

\item{X}{An \code{(nvars x nobs)} data frame or matrix in which each column contains 
all observations of measured (predictor) variables for a given timepoint or
sample.  For example, if X represents a spatial variable that was recorded 
at several timepoints, then each row of X should contain the variable's 
measurement for all timepoints at a single location.}
}
\value{
Y Predicted values for Y.
}
\description{
Make predictions using canonical correlation analysis results
}
\references{
Cook, E.R., Briffa, K.R., and Jones, P.D., 1994, Spatial regression methods in dendroclimatology: A review and comparison of two techniques: International Journal of Climatology, v. 14, p. 379–402.
}
\seealso{
\link{cca} For details about canonical correlation analyses.
}

