% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.stFit.R
\name{coef.stFit}
\alias{coef.stFit}
\title{Compute point estimates for parameters from posterior samples}
\usage{
\method{coef}{stFit}(stFit, burn = 1, fun = mean)
}
\arguments{
\item{stFit}{stFit object containing posterior samples for model}

\item{burn}{number of posterior samples to reject before computing estimates}

\item{fun}{function for computing point estimates}
}
\description{
Compute point estimates for parameters from posterior samples
}

