% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eof.R
\name{eof}
\alias{eof}
\title{Performs an EOF decomposition of the data}
\usage{
eof(X, center = F, scale = F)
}
\arguments{
\item{X}{[variable x observation] matrix of data for which to compute EOFs}
}
\value{
A list containing EOF patterns as columns, and their scores
}
\description{
Performs an EOF decomposition of the data
}
