% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRegion.R
\name{extractRegion}
\alias{extractRegion}
\title{Extract region from a SpatialGridDataFrame}
\usage{
extractRegion(sgdf, extent, type = "response", aggfact = NULL,
  mask = NULL)
}
\arguments{
\item{extent}{raster::extent object featuring region to extract}

\item{type}{whether to return the raw data, anomalies (data minus temporal 
average at each location), or standardized anomalies (anomalies divided by
temporal standard deviation at each location)}

\item{aggfact}{if provided, will spatially average the data}

\item{mask}{if an sgdf is provided, the data will be masked before
extraction, aggregation, and anomaly computation}
}
\value{
a modified SpatialGridDataFrame, sgdf, with the climatology for each
 location accessible via attr(sgdf@data@values, 'scaled:center') if anomalies
 were computed
}
\description{
Extract region from a SpatialGridDataFrame
}
\examples{
# Show an example where you plot the climatology grid too!


}

