% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maternEffectiveRange.R
\name{maternEffectiveRange}
\alias{maternEffectiveRange}
\title{Compute effective range for Matern correlation to drop to a specified level}
\usage{
maternEffectiveRange(cor = 0.05, range = 1, smoothness = 0.5)
}
\arguments{
\item{cor}{Effective correlation to check for}

\item{range}{Matern range parameter.  Controls the decay of pointwise 
correlations as a function of distance.}

\item{smoothness}{Matern smoothness parameter.  Controls the number of 
process derivatives.}
}
\description{
Compute effective range for Matern correlation to drop to a specified level
}

