% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.stData.R
\name{plot.stData}
\alias{plot.stData}
\title{Plot stData objects}
\usage{
plot.stData(stData, type = "response", t = NULL, boxsize = NULL,
  p = NULL, map = "world", region = ".", coord.s = NULL, zlim = NULL,
  lab.teleconnection = expression(alpha))
}
\arguments{
\item{stData}{Object of class stData to plot.}

\item{type}{Either 'response', 'covariate', 'remote', or 'teleconnection' 
to specify which part of stData to plot.  Note that 'teleconnection' applies
only if the stData object contains information about teleconnection effects,
i.e., if it is a simulated dataset or otherwise modified to include 
estimates of teleconnection effects.  Note that the value for type can
be an abbreviation since partial matching is used during plotting.}

\item{t}{timepoint to plot.  Will automatically plot the first timepoint if
t=NULL.}

\item{boxsize}{size of grid boxes plotted}

\item{p}{column index of local covariate to plot if type='covariate'. Will 
automatically assume the local covariate data includes an intercept and will
plot the second column if p=NULL.}

\item{map}{name of map provided by the maps package. These include county, 
france, italy, nz, state, usa, world, world2.  By default, all stData plots
will include us state outlines.}

\item{region}{name of subregions to include. Defaults to . which includes 
all subregions. See documentation for map for more details.}

\item{coord.s}{if plot type is 'teleconnection', specifies the longitude and 
latitude of local coordinate for which to plot teleconnection effects. if 
NULL, the middle local coordinate will be plotted.}

\item{zlim}{c(min, max) vector that specifies the colorscale limits}

\item{lab.teleconnection}{label used for fill scale in teleconnection plot}
}
\value{
a ggplot object with the specified map
}
\description{
This function provides basic plotting for telefit package data.
}

