% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.stFit.R
\name{plot.stFit}
\alias{plot.stFit}
\title{Plot stFit objects}
\usage{
plot.stFit(stFit, type = "density", stData = NULL, coord.s = NULL,
  text.size = NULL, axis.text.size = NULL, burn = 1, signif.telecon = F,
  p = 1, ...)
}
\arguments{
\item{stFit}{Object of class stFit to plot.}

\item{type}{Either 'traceplot', 'density', 'pairs', 'teleconnection',
'teleconnection_knot', or 'beta' to specify which part of stFit to plot. 
Note that the value for
type can be an abbreviation since partial matching is used during plotting.}

\item{stData}{Object of class stData to provide coordinate and related
information for plotting estimated teleconnection effects}

\item{coord.s}{if plot type is 'teleconnection', specifies the longitude and 
latitude of local coordinate for which to plot estimated teleconnection 
effects. if NULL, the middle local coordinate will be plotted.}

\item{text.size}{number specifying the size of text labels}

\item{axis.text.size}{number specifying the size of axis text labels}

\item{burn}{number of observations to exclude from graph}

\item{signif.telecon}{if TRUE, will highlight significant teleconnection
effects when type=='teleconnection'}

\item{p}{If stFit was fit with spatially varying coefficients, p specifies 
the index of the spatially varying coefficient to plot}
}
\value{
a ggplot object with the specified map
}
\description{
This function provides basic plotting for telefit package data.
}

