% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.stPredict.R
\name{plot.stPredict}
\alias{plot.stPredict}
\title{Plot stPredict objects}
\usage{
plot.stPredict(stPredict, type = "prediction", t = NULL, stFit = NULL,
  stData = NULL, err.comparison = NULL, err.var = NULL,
  err.lab = err.var, dots = NULL, ...)
}
\arguments{
\item{stPredict}{Object of class stPredict to plot.}

\item{type}{Either 'prediction', 'residual', 'observed', 'standard_error' (or 'se'), 
'local', 'remote', 'correlation', 'teleconnection', 'teleconnection_knot',
'teleconnection_knot_transect', 'errors', or 'cat.prediction'
to specify which part of stPredict to plot. Note that the value for type can
be an abbreviation since partial matching is used during plotting. 'truth'
and 'residual' plots are only available if the model has been evaluted and
the predictions have been compared to another response dataset.}

\item{t}{timepoint to plot.  Will automatically plot the first timepoint if
t=NULL.}

\item{stFit}{Object of class stFit to provide related
information and structures for plotting estimated teleconnection effects}

\item{stData}{Object of class stData to provide coordinate and related
information for plotting estimated teleconnection effects}

\item{err.comparison}{data.frame with Year column  and a column for a variable
that will be used to plot annual errors against}

\item{err.var}{name of variable in err.comparison for plotting against}

\item{err.lab}{label for name of variable in err.comparison for plotting against}

\item{...}{additional arguments to be passed to lower-level plotting functions}
}
\value{
a ggplot object with the specified map
}
\description{
This function provides basic plotting for telefit package data.
}

