% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.stPredict.R
\name{plot.stPredict}
\alias{plot.stPredict}
\title{Plot stPredict objects}
\usage{
plot.stPredict(stPredict, type = "prediction", t = NULL, stFit = NULL,
  stData = NULL, err.comparison = NULL, err.var = NULL,
  err.lab = err.var, pattern = 1, dots = NULL, burn = 1, ...)
}
\arguments{
\item{stPredict}{Object of class stPredict to plot.}

\item{type}{One of the following options to specify what type of plot to build
\describe{
  \item{prediction} { }
  \item{residual} { }
  \item{observed} { }
  \item{standard_error (or 'se')} { }
  \item{prediction} { }
  \item{local} { }
  \item{remote} { }
  \item{correlation} { }
  \item{teleconnection} { }
  \item{teleconnection_knot} { }
  \item{teleconnection_knot_transect} { }
  \item{errors} { }
  \item{cat.prediction} { }
  \item{truth} { Note: this plot is only available if the model has been
    evaluated and the predictions have been compared to another response
    dataset. }
  \item{residual} { Note: this plot is only available if the model has been
    evaluated and the predictions have been compared to another response
    dataset. }
  \item{eof-alpha_knots}{ A map of the local domain where the plotted colors
    show the remote influence coefficients mapped onto the eof pattern
    specified by the "pattern" argument.   }
}}

\item{t}{timepoint to plot.  Will automatically plot the first timepoint if
t=NULL.}

\item{stFit}{Object of class stFit to provide related
information and structures for plotting estimated teleconnection effects}

\item{stData}{Object of class stData to provide coordinate and related
information for plotting estimated teleconnection effects}

\item{err.comparison}{data.frame with Year column  and a column for a variable
that will be used to plot annual errors against}

\item{err.var}{name of variable in err.comparison for plotting against}

\item{err.lab}{label for name of variable in err.comparison for plotting against}

\item{pattern}{if type=='eof-alpha_knots', this specified which eof the remote 
coefficients should be mapped onto and then plotted over the local domain}

\item{burn}{number of observations to exclude from graph}

\item{...}{additional arguments to be passed to lower-level plotting functions}
}
\value{
a ggplot object with the specified map
}
\description{
This function provides basic plotting for telefit package data.
}

