% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.teleCor.R
\name{plot.teleCor}
\alias{plot.teleCor}
\title{Plots teleconnection correlation maps}
\usage{
plot.teleCor(teleCor, signif = F, coord.s = NULL, boxsize = NULL,
  map = "world", region = ".", zlim = NULL)
}
\arguments{
\item{teleCor}{object of class teleCor, containing pointwise correlations}

\item{signif}{if TRUE, then teleCor must have a column labeled 'signif' that
indicates which correlations are significant.  These correlations will be
printed in bold, and the rest will be printed more lightly}

\item{coord.s}{specifies the longitude and 
latitude of local coordinate for which to plot pointwise correlations. if 
NULL, the middle local coordinate will be plotted.}

\item{boxsize}{size of grid boxes plotted}

\item{map}{name of map provided by the maps package. These include county, 
france, italy, nz, state, usa, world, world2.  By default, all stData plots
will include us state outlines.}

\item{region}{name of subregions to include. Defaults to . which includes 
all subregions. See documentation for map for more details.}

\item{zlim}{c(min, max) vector that specifies the colorscale limits}
}
\value{
a ggplot object with the specified map
}
\description{
This function provides basic plotting for analyses returned from cor.tel
}

