% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stLL.R
\name{stLL}
\alias{stLL}
\title{Compute log likelihood for model}
\usage{
stLL(stData, stFit, beta, sigmasq_y, sigmasq_r, sigmasq_eps, rho_y, rho_r,
  X = stData$X, Y = stData$Y, Z = stData$Z, coords.s = stData$coords.s,
  coords.r = stData$coords.r, coords.knots = stFit$coords.knots,
  miles = T)
}
\arguments{
\item{stData}{Object with class 'stData' containing data needed to fit this 
model. The data need only be manually entered if not using a stData object.}

\item{stFit}{Object with class 'stFit' containing posterior parameter samples
needed to composition sample the teleconnection effects and generate 
posterior predictions. The data needed from stFit need only be manually 
entered if not using a stData object.}

\item{X}{[ns, p, nt] array of design matrices with local covariates}

\item{Y}{[ns, nt] matrix with response data}

\item{Z}{[nr, nt] matrix with remote covariates}

\item{coords.s}{matrix with coordinates where responses were 
observed (lon, lat)}

\item{coords.r}{matrix with coordinates where remote covariates
were observed (lon, lat)}

\item{stDataNew}{object of class stData that includes information needed for 
making forecasts.  If response data is included, this function will 
automatically run stEval using the empirical climatology as the reference
forecast}

\item{burn}{number of posterior samples to burn before drawing composition
samples}

\item{prob}{confidence level for approximate confidence intervals of 
teleconnection effects (only needed if returnAlphas==TRUE)}

\item{ncores}{Since the teleconnection effects and posterior predictions can 
be sampled in parallel, this parameter lets users specify the number of 
cores to use to draw teleconnection and prediction samples}

\item{returnAlphas}{TRUE to return the teleconnection effects sampled 
alongside the forecasts.  Note that only basic summary information about the 
teleconnection effects will be returned.}

\item{returnFullAlphas}{TRUE to return the teleconnection effects sampled 
alongside the forecasts.  Note that only basic summary information about the 
teleconnection effects will be returned.}

\item{conf}{Parameter specifying the HPD level to compute for posterior 
predictive samples}

\item{tLabs}{Forecast timepoint labels}

\item{Xnew}{[ns, p, nt0] array of design matrices with local covariates 
at forecast timepoints}

\item{Znew}{[nr, nt0] matrix with remote covariates at forecast timepoints}

\item{cat.probs}{vector of probabilities for also returning categorical 
predictions from the posterior prediction samples; NULL otherwise}
}
\description{
Compute log likelihood for model
}

