% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stPredict.R
\name{stPredict}
\alias{stPredict}
\title{Compute forecasts based on posterior samples}
\usage{
stPredict(stFit, stData, ncores = 1, conf = 0.95,
  coords.s = stData$coords.s, X = stData$X, Z = stData$Z,
  tLabs = stData$tLabs)
}
\arguments{
\item{stFit}{object of class stFit that includes estimated teleconnection 
effects.  This information will be used to make predictions}

\item{stData}{object of class stData that includes information needed for 
making forecasts.  If response data is included, this function will 
automatically run stEval using the empirical climatology as the reference
forecast}
}
\description{
Compute forecasts based on posterior samples
}

