% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stPredict.R
\name{stPredict}
\alias{stPredict}
\title{Compute forecasts based on posterior samples}
\usage{
stPredict(stFit, coords.local, X, Z, yearLabs, ncores = 1)
}
\description{
Compute forecasts based on posterior samples
}

