% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stSimulate.R
\name{stSimulate}
\alias{stSimulate}
\title{Simulate the spatio-temporal teleconnection model}
\usage{
stSimulate(X = NULL, Z = NULL, coords.s = NULL, coords.r = NULL,
  cov.s = NULL, cov.r = NULL, alpha = NULL, beta = NULL, nt = NULL,
  ns = NULL, nr = NULL, p = NULL, miles = T)
}
\description{
This function will let you provide as much information as you'd like, but 
will fill in the gaps as needed
}

