% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseAlpha.R
\name{summariseAlpha}
\alias{summariseAlpha}
\title{Summarize alphas}
\usage{
summariseAlpha(alpha, burn, prob = 0.95, coords.local, coords.remote)
}
\description{
This function computes approximate normal intervals, etc. for fitted alphas.
}

