% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseEOFAlpha.R
\name{summariseEOFAlpha}
\alias{summariseEOFAlpha}
\title{Summarize eof-mapped alphas}
\usage{
summariseEOFAlpha(eof_alpha, prob = 0.95, coords.s)
}
\description{
This function computes approximate normal intervals, etc. for fitted 
eof-mapped alphas.
}

