% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.stPredict.R
\name{summary.stPredict}
\alias{summary.stPredict}
\title{Plot stPredict objects}
\usage{
summary.stPredict(stPredict, t = NULL, digits = NULL)
}
\arguments{
\item{stPredict}{Object of class stPredict to summarise}

\item{t}{timepoint to plot.  Will automatically plot all timepoints and overall
summary if NULL.}

\item{digits}{Number of digits to pass to signif, if not NULL.}
}
\description{
This function prints basic summary info for telefit stPredict objects
}

