% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teleCor.R
\name{teleCor}
\alias{teleCor}
\title{Compute pointwise correlations for an exploratory teleconnection analysis}
\usage{
teleCor(stData = NULL, Y = stData$Y, Z = stData$Z,
  coords.s = stData$coords.s, coords.r = stData$coords.r, ncores = 1)
}
\arguments{
\item{Y}{[ny x nt]}

\item{Z}{[nz x nt]}

\item{coords.s}{coordinates of locations in Y}

\item{coords.r}{coordinates of locations in Z}
}
\value{
list with a matrix 'cor' containing correlations.  The columns index
 the remote coordinates, while the rows index the local coordinates
}
\description{
Compute pointwise correlations for an exploratory teleconnection analysis
}

