% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telefit.R
\docType{package}
\name{telefit}
\alias{telefit}
\alias{telefit-package}
\title{Tools for modeling teleconnections}
\description{
The package \pkg{telefit} provides functions for performing statistical methods taught in CSU's STAT 301 course. The functions offer easy input and convenient output for t-tests, ANOVA, and regression analysis.
}
